+++ Selectable.php
@@ -42,12 +42,45 @@ use yii\helpers\Html;
  * ]);
  * ```
  *
+ * Selectable in begin mode.
+ *
+ * ```php
+ * Selectable::begin([
+ *     'clientOptions' => [
+ *         'filter' => 'my-selectable-item',
+ *         'tolerance' => 'touch',
+ *     ],
+ * ]);
+ * ```
+ * <ul>
+ *      <li class="my-selectable-item">Item 1</li>
+ *      <li class="my-selectable-item">Item 2</li>
+ *      <li class="no-selectable-item">Item 3</li>
+ *      <li class="my-selectable-item">Item 4</li>
+ * </ul>
+ * <div>
+ *      <div>
+ *          <div class="my-selectable-item">Another item</div>
+ *      </div>
+ * </div>
+ *
+ * ```php
+ * Selectable::end();
+ * ```
+ *
  * @see http://api.jqueryui.com/selectable/
  * @author Alexander Kochetov <creocoder@gmail.com>
  * @since 2.0
  */
 class Selectable extends Widget
 {
+    const MODE_DEFAULT = 'MODE_DEFAULT';
+    const MODE_BEGIN = 'MODE_BEGIN';
+    
+    /**
+     * @var string the mode used to render the widget.
+     */
+    public $mode = self::MODE_DEFAULT;
     /**
      * @var array the HTML attributes for the widget container tag. The following special options are recognized:
      *
@@ -78,21 +111,53 @@ class Selectable extends Widget
      * @see \yii\helpers\Html::renderTagAttributes() for details on how attributes are being rendered.
      */
     public $itemOptions = [];
-
-
+    
+    /**
+     * Begins a widget.
+     * This method creates an instance of the calling class setting the MODE_BEGIN mode. Any item between 
+     * [[begin()]] and [[end()]] which match the filter attribute, will be selectable. 
+     * It will apply the configuration
+     * to the created instance. A matching [[end()]] call should be called later.
+     * As some widgets may use output buffering, the [[end()]] call should be made in the same view
+     * to avoid breaking the nesting of output buffers.
+     * @param array $config name-value pairs that will be used to initialize the object properties
+     * @return static the newly created widget instance
+     * @see end()
+     */
+    public static function begin($config = []) {
+        $config['mode'] = self::MODE_BEGIN;
+        parent::begin($config);
+    }
+    
+    /**
+     * Initializes the widget.
+     */
+    public function init()
+    {
+        parent::init();
+        if ($this->mode === self::MODE_BEGIN) {
+            echo Html::beginTag('div', $this->options) . "\n";
+        }
+    }
+    
     /**
      * Renders the widget.
      */
     public function run()
     {
-        $options = $this->options;
-        $tag = ArrayHelper::remove($options, 'tag', 'ul');
-        echo Html::beginTag($tag, $options) . "\n";
-        echo $this->renderItems() . "\n";
-        echo Html::endTag($tag) . "\n";
+        if ($this->mode === self::MODE_BEGIN) {
+            echo Html::endTag('div') . "\n";
+        } else {
+            $options = $this->options;
+            $tag = ArrayHelper::remove($options, 'tag', 'ul');
+            echo Html::beginTag($tag, $options) . "\n";
+            echo $this->renderItems() . "\n";
+            echo Html::endTag($tag) . "\n";
+        }
+        
         $this->registerWidget('selectable');
     }
-
+    
     /**
      * Renders selectable items as specified on [[items]].
      * @return string the rendering result.
@@ -115,7 +180,6 @@ class Selectable extends Widget
                 $items[] = Html::tag($tag, $item, $options);
             }
         }
-
         return implode("\n", $items);
     }
 }